Renamed MainCharMovement to Player
