+++ Assets/Scripts/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 29c6bc07a365cb542824a057f293750e
+guid: 26489197946ffd846a1e4e4267b7ed41
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
