+++ Assets/Utils/Singleton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 26489197946ffd846a1e4e4267b7ed41
+guid: 757e6b783a5c0c94490f015d5449db87
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
